[TOOLS-2426] Change Issue Kpi Filter location
