+++ kpi-parent/kpi-testutils/src/test/java/objective/taskboard/followup/kpi/IssueKpiServiceTest.java
@@ -116,194 +116,6 @@ public class IssueKpiServiceTest {
                 .amountOfIssueIs(0);
     }
 
-    @Test
-    public void getIssuesFromCurrentState_filteringOpen() {
-        KpiEnvironment context = dsl()
-                .services()
-                    .projects()
-                        .withKey("PROJ")
-                            .startAt("2020-01-04")
-                            .deliveredAt("2020-01-10")
-                        .eoP()
-                    .eoPs()
-                .eoS()
-                .todayIs("2020-01-12")
-                .givenFeature("I-1")
-                    .project("PROJ")
-                    .type("Task")
-                    .withTransitions()
-                        .status("Open").date("2020-01-01")
-                        .status("To Do").date("2020-01-02")
-                        .status("Doing").date("2020-01-03")
-                        .status("Done").noDate()
-                    .eoT()
-                    .subtask("I-2")
-                        .type("Alpha")
-                        .project("PROJ")
-                        .withTransitions()
-                            .status("Open").date("2020-01-01")
-                            .status("To Do").date("2020-01-02")
-                            .status("Doing").date("2020-01-03")
-                            .status("Done").date("2020-01-04")
-                        .eoT()
-                        .worklogs()
-                            .at("2020-01-03").timeSpentInHours(3.0)
-                        .eoW()
-                    .endOfSubtask()
-                    .subtask("I-3")
-                        .type("Alpha")
-                        .project("PROJ")
-                        .withTransitions()
-                            .status("Open").date("2020-01-01")
-                            .status("To Do").date("2020-01-02")
-                            .status("Doing").noDate()
-                            .status("Done").noDate()
-                        .eoT()
-                    .endOfSubtask()
-                    .subtask("I-4")
-                        .type("Alpha")
-                        .project("PROJ")
-                        .withTransitions()
-                            .status("Open").date("2020-01-01")
-                            .status("To Do").date("2020-01-02")
-                            .status("Doing").noDate()
-                            .status("Done").noDate()
-                        .eoT()
-                        .worklogs()
-                            .at("2020-01-03").timeSpentInHours(5.0)
-                        .eoW()
-                    .endOfSubtask()
-                .eoI();
-            context
-                .when()
-                    .appliesBehavior(getIssuesFromCurrentStateForProjectAndLevel("PROJ", FEATURES))
-                .then()
-                    .amountOfIssueIs(1)
-                    .givenIssue("I-1")
-                        .hasLevel(FEATURES)
-                        .hasType("Task")
-                        .atDate("2020-01-12").isOnStatus("Doing").eoDc()
-                        .atStatus("Doing").hasTotalEffortInHours(8.0).eoSa()
-                        .hasChild("I-2")
-                        .hasChild("I-3")
-                        .hasChild("I-4");
-            context
-                .when()
-                    .appliesBehavior(getIssuesFromCurrentStateForProjectAndLevel("PROJ", SUBTASKS))
-                .then()
-                    .amountOfIssueIs(2)
-                    .givenIssue("I-2")
-                        .hasLevel(SUBTASKS)
-                        .hasType("Alpha")
-                        .atDate("2020-01-12").isOnStatus("Done").eoDc()
-                        .atStatus("To Do").hasNoEffort().eoSa()
-                        .atStatus("Doing").hasTotalEffortInHours(3.0).eoSa()
-                        .atStatus("Done").hasNoEffort().eoSa()
-                    .eoIA()
-                    .givenIssue("I-4")
-                        .hasLevel(SUBTASKS)
-                        .hasType("Alpha")
-                        .atDate("2020-01-12").isOnStatus("To Do").eoDc()
-                        .atStatus("To Do").hasNoEffort().eoSa()
-                        .atStatus("Doing").hasTotalEffortInHours(5.0).eoSa()
-                        .atStatus("Done").hasNoEffort().eoSa()
-                    .eoIA();
-    }
-
-    @Test
-    public void getIssuesFromCurrentState_filteringClosedIssues() {
-        KpiEnvironment context = dsl()
-                .services()
-                    .projects()
-                        .withKey("PROJ")
-                            .startAt("2020-01-06")
-                            .deliveredAt("2020-01-10")
-                        .eoP()
-                    .eoPs()
-                .eoS()
-                .todayIs("2020-01-12")
-                .givenFeature("I-1")
-                    .project("PROJ")
-                    .type("Task")
-                    .withTransitions()
-                        .status("Open").date("2020-01-01")
-                        .status("To Do").date("2020-01-02")
-                        .status("Doing").date("2020-01-03")
-                        .status("Done").noDate()
-                    .eoT()
-                    .subtask("I-2")
-                        .type("Alpha")
-                        .project("PROJ")
-                        .withTransitions()
-                            .status("Open").date("2020-01-01")
-                            .status("To Do").date("2020-01-02")
-                            .status("Doing").date("2020-01-03")
-                            .status("Done").date("2020-01-04")
-                        .eoT()
-                        .worklogs()
-                            .at("2020-01-03").timeSpentInHours(3.0)
-                        .eoW()
-                    .endOfSubtask()
-                    .subtask("I-3")
-                        .type("Alpha")
-                        .project("PROJ")
-                        .withTransitions()
-                            .status("Open").date("2020-01-01")
-                            .status("To Do").date("2020-01-02")
-                            .status("Doing").date("2020-01-07")
-                            .status("Done").noDate()
-                        .eoT()
-                    .endOfSubtask()
-                    .subtask("I-4")
-                        .type("Alpha")
-                        .project("PROJ")
-                        .withTransitions()
-                            .status("Open").date("2020-01-01")
-                            .status("To Do").date("2020-01-02")
-                            .status("Doing").noDate()
-                            .status("Done").noDate()
-                        .eoT()
-                        .worklogs()
-                            .at("2020-01-09").timeSpentInHours(5.0)
-                        .eoW()
-                    .endOfSubtask()
-                .eoI();
-            context
-                .when()
-                    .appliesBehavior(getIssuesFromCurrentStateForProjectAndLevel("PROJ", FEATURES))
-                .then()
-                    .amountOfIssueIs(1)
-                    .givenIssue("I-1")
-                        .hasLevel(FEATURES)
-                        .hasType("Task")
-                        .atDate("2020-01-12").isOnStatus("Doing").eoDc()
-                        .atStatus("Doing").hasTotalEffortInHours(8.0).eoSa()
-                        .hasChild("I-2")
-                        .hasChild("I-3")
-                        .hasChild("I-4");
-            context
-                .when()
-                    .appliesBehavior(getIssuesFromCurrentStateForProjectAndLevel("PROJ", SUBTASKS))
-                .then()
-                    .amountOfIssueIs(2)
-                    .givenIssue("I-3")
-                        .hasLevel(SUBTASKS)
-                        .hasType("Alpha")
-                        .atDate("2020-01-12").isOnStatus("Doing").eoDc()
-                        .atStatus("To Do").hasNoEffort().eoSa()
-                        .atStatus("Doing").hasNoEffort().eoSa()
-                        .atStatus("Done").hasNoEffort().eoSa()
-                    .eoIA()
-                    .givenIssue("I-4")
-                        .hasLevel(SUBTASKS)
-                        .hasType("Alpha")
-                        .atDate("2020-01-12").isOnStatus("To Do").eoDc()
-                        .atStatus("To Do").hasNoEffort().eoSa()
-                        .atStatus("Doing").hasTotalEffortInHours(5.0).eoSa()
-                        .atStatus("Done").hasNoEffort().eoSa()
-                    .eoIA();
-    }
-
     private ServeIssuesFromDatasets getIssuesFromDataSet(KpiLevel kpiLevel) {
         return new ServeIssuesFromDatasets(kpiLevel);
     }
